Added strict keyword argument + switched ordering in step function
Previously, in every case when step() was called, results from futures was added at the beginning. This required a user to call step() again or some other function to update the data in xopt/generator. I've reversed it s.t. the futures are polled at the end of step() using the process_futures() method. In either ordering step() must wait for at least one future to finish (if asynch is true) or all of them to finish (if asynch is not true).

Step also raises an error if a future raises an error and strict=True